+++ CHANGELOG.md
@@ -18,6 +18,11 @@ The API consists of all public Kotlin types from `com.atlassian.performance.tool
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/ssh/branches/compare/master%0Drelease-2.0.0
 
+### Added
+- Support password authentication which resolves JPERF-237
+
+[JPERF-237]: https://ecosystem.atlassian.net/browse/JPERF-237
+
 ## [2.0.0] - 2018-10-26
 [2.0.0]: https://bitbucket.org/atlassian/ssh/branches/compare/release-2.0.0%0Drelease-1.2.0